Final Push

+++ assets/js/cs/task.jsx
@@ -6,24 +6,30 @@ import api from '../api';
 function Task(params) {
 
 	function edit(ev) {
-	    api.submit_task(props.form);
-	    console.log(props.form);
+	    let action = {
+	      type: 'UPDATE_EDITFORM',
+	      data: params.task,
+	    };
+	    params.dispatch(action);
 	}
 
 	function delete_task(ev) {
-	    api.delete_task(params.task.id);
+		var result = confirm("Are you sure?");
+		if(result){
+			api.delete_task(params.task.id);
+		}	    
 	}
 
   let task = params.task;
   return <Card>
     <CardBody>
       <div>
-        <p><b>Assigned To:</b> { task.user.name } </p>
+        <p><b>Assigned To:</b> { task.name } </p>
         <p><b>Title:</b> { task.title }</p>
         <p><b>Description:</b> { task.description }</p>
         <p><b>Time Spent:</b> { task.time } minutes</p>
         <p><b>Completed:</b> { task.completed } </p>
-        <Button onClick={delete_task} color="primary">Delete</Button> &nbsp;
+        <Button className="btn btn-default btn-danger" onClick={delete_task}>Delete</Button> &nbsp;
         <Button onClick={edit} color="primary">Edit</Button> &nbsp;
       </div>
     </CardBody>
@@ -31,9 +37,8 @@ function Task(params) {
 }
 
 function state2props(state) {
-  console.log("rerender@TaskForm", state);
   return {
-    form: state.form,
+    editform: state.editform,
     users: state.users,
   };
 }
